Rename scripts/i18n-filter.ts to translator.ts
