+++ src/generic_ui/polymer/i18n-filter.ts
@@ -1,7 +1,7 @@
 /// <reference path='../../../../third_party/polymer/polymer.d.ts' />
 
-import i18n_module = require('../scripts/i18n-filter');
-import i18n_t = i18n_module.i18n_t;
+import translator_module = require('../scripts/translator');
+import i18n_t = translator_module.i18n_t;
 
- declare var PolymerExpressions: any;
- PolymerExpressions.prototype.$$ = i18n_t;
+declare var PolymerExpressions: any;
+PolymerExpressions.prototype.$$ = i18n_t;
