+++ src/generic_ui/scripts/translator.ts
